requirements.txt
google-adk>=1.0.0
litellm>=1.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic-settings>=2.7.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-gcp-trace>=1.6.0
google-cloud-aiplatform[agent_engines,adk]>=1.76.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
httpx>=0.27.0
